Queue System Finished and Bug Fixes
Queue System:
Create a VC named "Queue" and VC that has "| Among Us" somewhere in it.

Join the Queue, you should be dragged to the game channel.

If there are more than 10 people in game vc you wont be dragged. When someone leaves the game vc while your in the queue vc you will be dragged (based on who has been in queue the longest)
-

- Line 57 no longer throws and error if there's no channel that matches

- Support for 6 digit codes (for Among Us test branch )

- Todo list to readme